src/requirements.txt
Django==2.2.7
djangorestframework==3.11.0
psycopg2-binary==2.8.4
drf-yasg==1.17.0